requirements.txt
Django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.3
google-genai>=0.3.0
python-dotenv>=1.0
